Fix a client-side bug for named books with no pages.
Changing the name or lore of a book with no "pages" NBT would cause the
client to incorrectly believe that the last page is not writable,
regardless of the page count.

This would correct itself after the first time opening the book. We will
simulate opening and saving the book to fix this bug.