Fix creative mode for 1.6.2 - reset item name with a hack.
In creative mode, the client sends back the entire item stack (for no
good reason), replacing the one on the server side. ItemRenamer used 
to store the original unchanged version as a custom NBT tag, but in
1.6.1 and 1.6.2, Bukkit started to remove any raw tags during packet 
reading.

To overcome this, we store the NBT compound as "character codes" in 
the display name.