Fix signed/edited books retaining their old name in creative mode.
A Book & Quil renamed with ItemRenamer will not be properly renamed to
the title when signed in creative mode, and retain its old name. We 
simply forgot to intercept certain packets (15 and 250), and undo the 
renaming before it gets accepted on the server side.

This bug was discovered by pilvimaa. :)